requirements.txt
Dependencies to Install:
ultralytics
opencv-python
pyserial